Style and simplify symptom log screen (#521)
Why:
We would like to update the styles and layout of the symptom log screen.

This commit:
Moves the create symptom log button to the bottom of the screen. We also
refactored the list to be flatter, previously we were rendering lists of
lists to group the symptom logs by date. By rendering the full date on
each symptom log entry, we were able to reduce the complexity in the
codebase.
