+++ src/MyHealth/symptoms.ts
@@ -1,4 +1,4 @@
-import { Posix, beginningOfDay } from "../utils/dateTime"
+import { Posix } from "../utils/dateTime"
 
 export enum HealthAssessment {
   AtRisk,
@@ -29,8 +29,6 @@ export type SymptomLogEntry = {
 
 export type SymptomLogEntryAttributes = Omit<SymptomLogEntry, "id">
 
-type LogDataByDay = Record<Posix, SymptomLogEntry[]>
-
 export type DayLogData = {
   date: Posix
   symptomLogEntries: SymptomLogEntry[]
@@ -43,54 +41,4 @@ export const determineHealthAssessment = (
     return HealthAssessment.AtRisk
   }
   return HealthAssessment.NotAtRisk
-}
-
-type WithADate = { date: Posix }
-const compareDatesAscending = (
-  { date: dateLeft }: WithADate,
-  { date: dateRight }: WithADate,
-) => {
-  return dateLeft < dateRight ? -1 : 1
-}
-const compareDatesDescending = (
-  { date: dateLeft }: WithADate,
-  { date: dateRight }: WithADate,
-) => {
-  return dateLeft < dateRight ? 1 : -1
-}
-
-const logEntryBuckets = (allEntries: SymptomLogEntry[]): LogDataByDay => {
-  const groupedEntries: LogDataByDay = {}
-
-  allEntries.forEach((entry) => {
-    const { date } = entry
-    const entryDateBeginningOfDay = beginningOfDay(date)
-
-    if (groupedEntries[entryDateBeginningOfDay]) {
-      const newLogEntries = groupedEntries[entryDateBeginningOfDay]
-      newLogEntries.push(entry)
-      newLogEntries.sort(compareDatesAscending)
-      groupedEntries[entryDateBeginningOfDay] = newLogEntries
-    } else {
-      groupedEntries[entryDateBeginningOfDay] = [entry]
-    }
-  })
-
-  return groupedEntries
-}
-
-export const dayLogData = (
-  symptomLogEntries: SymptomLogEntry[],
-): DayLogData[] => {
-  const symptomEntriesLogData = logEntryBuckets(symptomLogEntries)
-
-  return Object.keys(symptomEntriesLogData)
-    .map((date: string) => {
-      const posixDate = Number(date)
-      return {
-        date: posixDate,
-        symptomLogEntries: symptomEntriesLogData[posixDate],
-      }
-    })
-    .sort(compareDatesDescending)
 }
