+++ src/factories/symptomLogContext.ts
@@ -2,7 +2,7 @@ import { Factory } from "fishery"
 import { SymptomLogState } from "../MyHealth/SymptomLogContext"
 
 export default Factory.define<SymptomLogState>(() => ({
-  dailyLogData: [],
+  symptomLogEntries: [],
   addLogEntry: jest.fn(),
   updateLogEntry: jest.fn(),
   deleteLogEntry: jest.fn(),
